+++ src/controllers/campaigns/CampaignsController.php
@@ -27,12 +27,13 @@ class CampaignsController extends Controller
      * Index
      *
      * @param array $variables
+     * @return Response
      */
-    public function actionIndex(array $variables = [])
+    public function actionIndex(array $variables = []): Response
     {
         $variables = [];
 
-        QARR::$plugin->routeTemplate('campaigns/index', $variables);
+        return $this->renderTemplate('qarr/campaigns/index', $variables);
     }
 
     // Private Methods