+++ src/web/assets/Dashboard.php
@@ -3,6 +3,7 @@
 namespace owldesign\qarr\web\assets;
 
 use craft\web\AssetBundle;
+use craft\web\assets\cp\CpAsset;
 
 class Dashboard extends AssetBundle
 {
@@ -17,6 +18,7 @@ class Dashboard extends AssetBundle
         $this->sourcePath = __DIR__;
 
         $this->depends = [
+            CpAsset::class,
             Tippy::class
         ];